Add empty feature repo
