+++ sdk/python/tests/example_repos/empty_feature_repo.py
@@ -1,0 +1,3 @@
+# This example feature repo is deliberately left empty. It should be used for tests that do not need
+# any feature views or other objects (for example, a test that checks that a feature service can be
+# applied and retrieved correctly).
