+++ sdk/python/tests/integration/e2e/test_universal_e2e.py
@@ -2,12 +2,8 @@ from datetime import timedelta
 
 import pytest
 
-from feast import BigQuerySource, Entity, FeatureView, Field
-from feast.types import Float32, String
 from tests.integration.feature_repos.universal.entities import driver
 from tests.integration.feature_repos.universal.feature_views import driver_feature_view
-from tests.utils.basic_read_write_test import basic_rw_test
-from tests.utils.cli_repo_creator import CliRunner, get_example_repo
 from tests.utils.e2e_test_validation import validate_offline_online_store_consistency
 
 
@@ -31,43 +27,3 @@ def test_e2e_consistency(environment, e2e_data_sources, infer_features):
     split_dt = df["ts_1"][4].to_pydatetime() - timedelta(seconds=1)
 
     validate_offline_online_store_consistency(fs, fv, split_dt)
-
-
-@pytest.mark.integration
-def test_partial() -> None:
-    """
-    Add another table to existing repo using partial apply API. Make sure both the table
-    applied via CLI apply and the new table are passing RW test.
-    """
-
-    runner = CliRunner()
-    with runner.local_repo(
-        get_example_repo("example_feature_repo_1.py"), "bigquery"
-    ) as store:
-        driver = Entity(name="driver", join_keys=["test"])
-
-        driver_locations_source = BigQuerySource(
-            table="feast-oss.public.drivers",
-            timestamp_field="event_timestamp",
-            created_timestamp_column="created_timestamp",
-        )
-
-        driver_locations_100 = FeatureView(
-            name="driver_locations_100",
-            entities=[driver],
-            ttl=timedelta(days=1),
-            schema=[
-                Field(name="lat", dtype=Float32),
-                Field(name="lon", dtype=String),
-                Field(name="name", dtype=String),
-                Field(name="test", dtype=String),
-            ],
-            online=True,
-            batch_source=driver_locations_source,
-            tags={},
-        )
-
-        store.apply([driver_locations_100])
-
-        basic_rw_test(store, view_name="driver_locations")
-        basic_rw_test(store, view_name="driver_locations_100")
