+++ sdk/python/tests/unit/local_feast_tests/test_feature_service_apply.py
@@ -10,7 +10,6 @@ def test_read_pre_applied() -> None:
     with runner.local_repo(
         get_example_repo("example_feature_repo_1.py"), "file"
     ) as store:
-
         assert len(store.list_feature_services()) == 1
         fs = store.get_feature_service("driver_locations_service")
         assert len(fs.tags) == 1
